requirements.txt
discord.py[voice]==1.6.0
PyQt5==5.15.1
sounddevice==0.4.1
